backend/requirements.txt
django
mssql-django
pyodbc
django-rest-framework
python-dotenv